% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdchecklist.R
\name{ebirdchecklist}
\alias{ebirdchecklist}
\title{View Checklist}
\usage{
ebirdchecklist(subId, sleep = 0, key = NULL, ...)
}
\arguments{
\item{subId}{The checklist identifier}

\item{sleep}{Time (in seconds) before function sends API call (defaults to
zero. Set to higher number if you are using this function in a loop with
many API calls).}

\item{key}{eBird API key. You can obtain one from https://ebird.org/api/keygen.
We strongly recommend storing it in your \code{.Renviron} file as an
environment variable called \code{EBIRD_KEY}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A data.frame containing:
}
\description{
View Checklist
}
\examples{
\dontrun{
ebirdchecklist("S121423354")
}
}
\references{
\url{http://ebird.org/}
}
