% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdhotspotinfo.R
\name{ebirdhotspotinfo}
\alias{ebirdhotspotinfo}
\title{Hotspot info}
\usage{
ebirdhotspotinfo(locId, key = NULL, ...)
}
\arguments{
\item{locId}{The hotspot to be checked. A single hotspot only.}

\item{key}{ebird API key. You can obtain one from https://ebird.org/api/keygen.
We strongly recommend storing it in your \code{.Renviron} file as an 
environment variable called \code{EBIRD_KEY}.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A data frame containing:

"locId":             The hotspot ID

"name":              The name of the hotspot

"latitude":          hotspot latitude

"longitude":         hotspot longitude

"countryCode":       hostpot country code

"countryName":       country name

"subnational1Name":  Subnational name - e.g. state name

"subnational1Code":  Subnational code e.g state code

"subnational2Code":  Subnational 2 code e.g. county code

"subnational2Name":  Subnational 2 code e.g. county name

"isHotspot": true,   true if a hotspot

"hierarchical.name":  Composite of name beginning with country name

"locID": "L99381"    The hotspot ID
}
\description{
Hotspot info
}
\examples{
\dontrun{
ebirdhotspotinfo("L99381")
}
}
\references{
\url{http://ebird.org/}
}
\author{
Guy Babineau \email{guy.babineauo@gmail.com},
}
