% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequencies.R
\name{ebirdloccheck}
\alias{ebirdloccheck}
\title{Check if a location type is valid}
\usage{
ebirdloccheck(loctype, loc)
}
\arguments{
\item{loctype}{one of: 'country', 'states', 'counties'}

\item{loc}{the location code. See \link{ebirdfreq} documentation for examples}
}
\description{
Check if a location type is valid
}
\keyword{internal}

