% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdtaxonomyversion.R
\name{ebirdtaxonomyversion}
\alias{ebirdtaxonomyversion}
\title{eBird Taxonomy Version}
\usage{
ebirdtaxonomyversion(key = NULL, ...)
}
\arguments{
\item{key}{optional eBird API key. You can obtain one from https://ebird.org/api/keygen.
We strongly recommend storing it in your \code{.Renviron} file as an
environment variable called \code{EBIRD_KEY} to avoid having to constantly
supply the key, and to avoid accidentally sharing it publicly.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
data.frame containing the collected information:

"authorityVer": Character of version.

"latest": Boolean indicating whether `authorityVer`` is the latest taxonomy version
}
\description{
Returns a data.frame of available version numbers of the eBird taxonomy
}
\examples{
\dontrun{
ebirdtaxonomyversion()
}
}
\references{
\url{http://ebird.org/}
}
\author{
Jordan Bradford \email{jrdnbradford@gmail.com}
}
