% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/frequencies.R
\name{eloc_check}
\alias{eloc_check}
\title{Check if a location type is valid}
\usage{
eloc_check(loctype, loc)
}
\arguments{
\item{loctype}{one of: 'country', 'states', 'counties'}

\item{loc}{the location code. See \link{get_freq} documentation for examples}
}
\description{
Check if a location type is valid
}
\keyword{internal}

