% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebird-defunct.R
\name{rebird-defunct}
\alias{rebird-defunct}
\title{Defunct functions in rebird}
\description{
\itemize{
 \item \code{\link{ebirdregioncheck}}: Use 'ebirdregioninfo' instead.
 \item \code{\link{ebirdloc}}: Use 'ebirdregion' instead.
 \item \code{\link{ebirdhotspot}}: Use 'ebirdregion' instead.
}
}
