% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebird-deprecated.R
\name{rebird-deprecated}
\alias{rebird-deprecated}
\title{Deprecated functions in rebird}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{ebirdregioncheck}}: Deprecated: 'ebirdregioncheck' will be removed in the next version of rebird. Use 'ebirdregioninfo' instead.
 \item \code{\link{ebirdloc}}: TDeprecated: 'ebirdloc' will be removed in the next version of rebird as it might not be suported in the new eBird API. Use 'ebirdregion' instead.
 \item \code{\link{ebirdhotspot}}: Deprecated: 'ebirdhotspot' will be removed in the next version of rebird as it might not be suported in the new eBird API. Use 'ebirdregion' instead.
}
}
