% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_code.R
\name{species_code}
\alias{species_code}
\title{Return species code}
\usage{
species_code(sciname = NULL)
}
\arguments{
\item{sciname}{(required) Character string of length 1 with the scientific name to
look for. Case insensitive.}
}
\value{
A character string with the eBird species code.
}
\description{
Returns the species code for a given scientific name. Uses an internally-stored
 version of the taxonomy. Also provides a message with the common name, scientific
 name, and species code of the species.
}
\examples{
species_code("Anhinga anhinga")
}
\references{
\url{http://ebird.org/}
}
\author{
Sebastian Pardo \email{sebpardo@gmail.com}
}
