% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax}
\alias{tax}
\title{2023 edition of the eBird taxonomy (eBird v2023)}
\format{
## `tax`
A data frame with 17,064 rows and 15 columns:
\describe{
  \item{sciName}{Scientific name}
  \item{comName}{Common name}
  \item{speciesCode}{Species code}
  \item{category}{Taxon category}
  \item{taxonOrder}{Taxonomic order}
  \item{bandingCodes}{Banding lookup codes}
  \item{comNameCodes}{Common name lookup codes}
  \item{sciNameCodes}{Scientific name lookup codes}
  \item{order}{Order scientific name}
  \item{familyCode}{Family code}
  \item{familyComName}{Family common name}
  \item{familySciName}{Common name}
  \item{reportAs}{Report as taxon}
  \item{extinct}{Extinct}
  \item{extinctYear}{Extinction year}
}
}
\source{
<https://www.birds.cornell.edu/clementschecklist/download/>
}
\usage{
tax
}
\description{
Clements, J. F., P. C. Rasmussen, T. S. Schulenberg, M. J. Iliff, 
T. A. Fredericks, J. A. Gerbracht, D. Lepage, A. Spencer, S. M. Billerman, 
B. L. Sullivan, and C. L. Wood. 2023. The eBird/Clements checklist of Birds 
of the World: v2023. Downloaded from 
\url{https://www.birds.cornell.edu/clementschecklist/download/}
}
\keyword{datasets}
