% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getlatlng.R
\name{getlatlng}
\alias{getlatlng}
\title{get latitude and longitude from ip address}
\usage{
getlatlng()
}
\value{
a vector of length 2 with lat, lng in that order
}
\description{
Returns the most recent and nearest reported sighting information
with observations of a species.
}
\examples{
\dontrun{
getlatlng()
}
}
\author{
Andy Teucher \email{andy.teucher@gmail.com}
}
\references{
\url{http://ipinfo.io}
}

