% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rebird-package.R
\docType{package}
\name{rebird-package}
\alias{rebird}
\alias{rebird-package}
\title{rebird: an R package for access to the eBird database.}
\description{
An R package to interface with the eBird database.
}
\details{
\tabular{ll}{
Package: \tab rebird\cr
Type: \tab Package\cr
Version: \tab 0.1.2.99\cr
Date: \tab 2014-10-31\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr}

This package provides an R interface to the recent eBird database.
}
\examples{
#see functions.
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu},
Scott Chamberlain \email{myrmecocystus@gmail.com}, Andy Teucher \email{andy.teucher@gmail.com}
}
\keyword{package}

