\docType{package}
\name{rebird-package}
\alias{rebird}
\alias{rebird-package}
\title{rEbird: an R package for access to the eBird database.}
\description{
  An R package to interface with the eBird webservices.
}
\details{
  \tabular{ll}{ Package: \tab reBird\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2012-10-13\cr
  License: \tab BSD\cr LazyLoad: \tab yes\cr}

  This package provides an R interface to the recent eBird
  database.
}
\examples{
#see functions.
}
\author{
  Rafael Maia \email{rm72@zips.uakron.edu}, Scott
  Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

